\name{notdocumentedyet}
\alias{notdocumentedyet}
%
%
%
%
%
\alias{rainfall}
\alias{gala}
%
%
%
%
% 20110321; misc. datasets.
\alias{fibre1.5}
\alias{fibre15}
%
%
%
%
%
%
%
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Undocumented Data Sets }
\description{
  Those currently undocumented and internally used functions are aliased
  to this help file.
  Ditto for some classes.

}
%\usage{
%normal1(lmean = "identity", lsd = "loge", zero = NULL)
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%  \item{lmean}{
%  Link function applied to the mean.
%  See \code{\link{Links}} for more choices.
%
%  }
%}
\details{
  In the \pkg{VGAMdata} package there are currently some
  objects which are undocumented.
  The help file suppresses the warnings when the package is 'CHECK'ed.


}
\value{
  Each objects may or may not have its own individual value.
  These will be documented over time.


}
%\references{
%}

\author{ T. W. Yee }
%\note{
%
%}
%\seealso{
%  \code{gaussianff},
%  \code{\link{posnormal1}}.
%}
%\examples{
%}
\keyword{models}
\keyword{regression}
\keyword{internal}







