\name{notdocumentedyet}
\alias{notdocumentedyet}
%
%
%
% 202409
\alias{adjust0.loglaplace1}
\alias{adjust01.logitlaplace1}
\alias{alaplace1.control}
\alias{alaplace2.control}
\alias{alaplace3.control}
\alias{dlogitlap}
\alias{dloglap}
\alias{logitlaplace1.control}
\alias{loglaplace1.control}
\alias{plogitlap}
\alias{ploglap}
\alias{qlogitlap}
\alias{qloglap}
\alias{rlogitlap}
\alias{rloglap}
% 202011
\alias{dgenpois}
% 20200825
\alias{prob.munb.size.VGAM}
%
%
%
% 20200622
\alias{deflat.limit.oipospois}
\alias{deflat.limit.oizeta}
%
%
%
%
%
%
\alias{rainfall}
\alias{gala}
%
%
%
%
% 20110321; misc. datasets.
\alias{fibre1.5}
\alias{fibre15}
%
%
%
%
%
%
%
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Undocumented Data Sets and Functions }
\description{
  Those currently undocumented and internally used functions are
  aliased to this help file.
  Ditto for some classes.

}
%\usage{
%normal1(lmean = "identity", lsd = "loge", zero = NULL)
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%  \item{lmean}{
%  Link function applied to the mean.
%  See \code{\link[VGAM}{Links}} for more choices.
%
%  }
%}
\details{
  In the \pkg{VGAMdata} package there are currently some
  objects which are undocumented.
  The help file suppresses the warnings when the package is 'CHECK'ed.


}
\value{
  Each objects may or may not have its own individual value.
  These will be documented over time.


}
%\references{
%}

\author{ T. W. Yee }
%\note{
%
%}
%\seealso{
%  \code{gaussianff},
%  \code{\link[VGAM}{posnormal1}}.
%}
%\examples{
%}
\keyword{models}
\keyword{regression}
\keyword{internal}







